hustlr-ml/requirements.txt
scikit-learn>=1.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
joblib>=1.3.0
pydantic>=2.6.0
supabase>=2.3.0
python-dotenv>=1.0.0
prophet>=1.1.5
requests>=2.25.0
torch>=2.0.0
torch-geometric>=2.3.0
